requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==2.11.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
nltk==3.8.1
numpy==1.24.3
openai==1.75.0
python-multipart==0.0.5
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.2.3
beautifulsoup4==4.13.4
markdown==3.8
scikit-learn==1.3.0
rouge-score==0.1.2
